PDF to audio reader
This code helps you automatically read any pdf file in English and saves it in a mp3 file. If you are driving a car or have no means to read it yourself, it is a good way to convert it into an audio file and listen to it on the way.